requirements.txt
absl-py==1.0.0
altgraph==0.17.2
appdirs==1.4.4
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
astroid==2.11.3
asttokens==2.0.5
astunparse==1.6.3
async-generator==1.10
atomicwrites==1.4.0
attrs==21.4.0
autopep8==1.6.0
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.0
bs4==0.0.1
cachetools==5.0.0
camelot-py==0.10.1
certifi==2021.10.8
cffi==1.15.0
chardet==3.0.4
charset-normalizer==2.0.12
ci-info==0.2.0
click==7.1.2
colorama==0.4.4
configobj==5.0.6
configparser==5.2.0
crypto==1.4.1
cryptography==37.0.0
cssselect==1.1.0
cycler==0.11.0
debugpy==1.6.0
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.4
distlib==0.3.4
distro==1.7.0
document==1.0
Elixir==0.7.1
entrypoints==0.4
et-xmlfile==1.1.0
etelemetry==0.3.0
executing==0.8.3
fastjsonschema==2.15.3
filelock==3.6.0
fitz==0.0.1.dev2
flatbuffers==1.12
fonttools==4.33.3
future==0.18.2
gast==0.4.0
ghostscript==0.7
google-auth==2.6.6
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
greenlet==1.1.2
grpcio==1.45.0
h11==0.13.0
h5py==3.6.0
httplib2==0.20.4
idna==2.10
importlib-metadata==4.11.3
importlib-resources==5.7.1
iniconfig==1.1.1
ipykernel==6.13.0
ipython==8.2.0
ipython-genutils==0.2.0
ipywidgets==7.7.0
isodate==0.6.1
isort==5.10.1
itchat==1.3.10
jdcal==1.4.1
jedi==0.18.1
jieba==0.42.1
Jinja2==3.1.1
joblib==1.1.0
jsonschema==4.4.0
jupyter==1.0.0
jupyter-client==7.3.0
jupyter-console==6.4.3
jupyter-core==4.10.0
jupyterlab-pygments==0.2.2
jupyterlab-widgets==1.1.0
keras==2.9.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.2
lazy-object-proxy==1.7.1
libclang==14.0.1
lxml==4.8.0
Markdown==3.3.6
MarkupSafe==2.1.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
mccabe==0.7.0
mistune==0.8.4
more-itertools==8.12.0
mysql-connector==2.2.9
Naked==0.1.31
nbclient==0.6.0
nbconvert==6.5.0
nbformat==5.3.0
nest-asyncio==1.5.5
networkx==2.8
neurdflib==5.0.1
nibabel==3.2.2
nipype==1.7.1
notebook==6.4.11
numpy==1.22.3
oauthlib==3.2.0
opencv-python==4.5.5.64
openpyxl==3.0.9
opt-einsum==3.3.0
outcome==1.1.0
packaging==21.3
pandas==1.4.2
pandocfilters==1.5.0
parsel==1.6.0
parso==0.8.3
pathlib==1.0.1
pbr==5.8.1
pdf2image==1.16.0
pdfminer==20191125
pdfminer.six==20220319
pdfplumber==0.6.1
pdftopng==0.2.3
pefile==2021.9.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.1.0
platformdirs==2.5.2
pluggy==1.0.0
prometheus-client==0.14.1
prompt-toolkit==3.0.29
protobuf==3.19.4
prov==2.0.0
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pycparser==2.21
pycryptodome==3.14.1
pydot==1.4.2
pydotplus==2.0.2
Pygments==2.12.0
pyinstaller==5.0.1
pyinstaller-hooks-contrib==2022.4
pylint==2.13.7
PyMuPDF==1.19.6
PyMySQL==1.0.2
pyOpenSSL==22.0.0
pyparsing==3.0.8
pyPdf==1.13
PyPDF2==1.27.9
pypng==0.0.21
PyQRCode==1.2.1
PyQt5==5.15.4
pyqt5-plugins==5.15.4.2.2
PyQt5-Qt5==5.15.2
PyQt5-sip==12.10.1
PyQt5-stubs==5.15.6.0
pyqt5-tools==5.15.4.3.2
pyrsistent==0.18.1
PySide2==5.15.2.1
PySocks==1.7.1
pytest==7.1.2
python-dateutil==2.8.2
python-docx==0.8.11
python-dotenv==0.20.0
pytz==2022.1
pywin32==303
pywin32-ctypes==0.2.0
pywinpty==2.0.5
pyxnat==1.4
PyYAML==6.0
pyzmq==22.3.0
qt5-applications==5.15.2.2.2
qt5-tools==5.15.2.1.2
qtconsole==5.3.0
QtPy==2.0.1
rdflib==6.1.1
requestium==0.1.13
requests==2.27.1
requests-file==1.5.1
requests-oauthlib==1.3.1
rsa==4.8
scikit-learn==1.0.2
scipy==1.8.0
seaborn==0.11.2
selenium==4.1.3
Send2Trash==1.8.0
shellescape==3.8.1
shiboken2==5.15.2.1
simplejson==3.17.6
six==1.16.0
sklearn==0.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.36
stack-data==0.2.0
stevedore==3.5.0
tabula-py==2.3.0
tabulate==0.8.9
Tempita==0.5.2
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.9.1
tensorflow-estimator==2.9.0
tensorflow-io-gcs-filesystem==0.25.0
termcolor==1.1.0
terminado==0.13.3
testpath==0.6.0
tf-estimator-nightly==2.8.0.dev2021122109
thop==0.0.31.post2005241907
threadpoolctl==3.1.0
tinycss2==1.1.1
tldextract==3.2.1
toml==0.10.2
tomli==2.0.1
torch==1.11.0
torchvision==0.12.0
tornado==6.1
tqdm==4.64.0
traitlets==5.1.1
traits==6.3.2
trio==0.20.0
trio-websocket==0.9.2
typed-ast==1.5.3
typing_extensions==4.2.0
Unidecode==1.3.4
urllib3==1.26.9
virtualenv==20.14.1
virtualenv-clone==0.5.7
virtualenvwrapper==4.8.4
w3lib==1.22.0
Wand==0.6.7
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.1.1
widgetsnbextension==3.6.0
wrapt==1.14.0
wsproto==1.1.0
xlrd==2.0.1
XlsxWriter==3.0.3
xlwt==1.3.0
zhdate==0.1
zipp==3.8.0
